+++ packages/localizations/src/ca-ES.ts
@@ -148,7 +148,7 @@ export const caES: LocalizationResource = {
       },
     },
     paymentMethods__label: 'Mètodes de pagament',
-    popular: 'Popular',
+    highlightedPlanBadge: 'Popular',
     pricingTable: {
       billingCycle: 'Cicle de facturació',
       included: 'Inclòs',
@@ -193,6 +193,20 @@ export const caES: LocalizationResource = {
     },
     navbar: {
       title: "Configura l'inici de sessió únic (SSO)",
+    },
+    selectProviderStep: {
+      title: 'Seleccioneu un proveïdor',
+      subtitle: 'Seleccioneu el proveïdor per al qual configurareu SSO.',
+      body: {
+        title: "Seleccioneu el vostre proveïdor d'identitat",
+        description: 'Us guiarem pel procés de configuració detallat a continuació.',
+      },
+      saml: {
+        groupLabel: 'SAML',
+        okta: 'Okta Workforce',
+        customSaml: 'Proveïdor SAML personalitzat',
+      },
+      warning: 'Un cop seleccionat un proveïdor no podreu canviar-lo fins que la configuració hagi finalitzat',
     },
     verifyEmailDomainStep: {
       title: 'Verifica el correu electrònic',