+++ packages/localizations/src/es-ES.ts
@@ -147,7 +147,7 @@ export const esES: LocalizationResource = {
       },
     },
     paymentMethods__label: 'Métodos de pago',
-    popular: 'Popular',
+    highlightedPlanBadge: 'Popular',
     pricingTable: {
       billingCycle: 'Ciclo de facturación',
       included: 'Incluido',
@@ -192,6 +192,20 @@ export const esES: LocalizationResource = {
     },
     navbar: {
       title: 'Configurar inicio de sesión único (SSO)',
+    },
+    selectProviderStep: {
+      title: 'Seleccionar proveedor',
+      subtitle: 'Selecciona el proveedor para el que vas a configurar SSO.',
+      body: {
+        title: 'Selecciona tu proveedor de identidad',
+        description: 'Te guiaremos a través del proceso de configuración detallado a continuación.',
+      },
+      saml: {
+        groupLabel: 'SAML',
+        okta: 'Okta Workforce',
+        customSaml: 'Proveedor SAML personalizado',
+      },
+      warning: 'Una vez seleccionado un proveedor no podrás cambiarlo hasta que finalice la configuración',
     },
     verifyEmailDomainStep: {
       title: 'Verificar correo electrónico',