+++ packages/localizations/src/kk-KZ.ts
@@ -141,7 +141,7 @@ export const kkKZ: LocalizationResource = {
       },
     },
     paymentMethods__label: undefined,
-    popular: undefined,
+    highlightedPlanBadge: undefined,
     pricingTable: {
       billingCycle: undefined,
       included: undefined,
@@ -185,6 +185,20 @@ export const kkKZ: LocalizationResource = {
     },
     navbar: {
       title: 'Бірыңғай кіруді конфигурациялау (SSO)',
+    },
+    selectProviderStep: {
+      title: 'Select provider',
+      subtitle: 'Select the provider you are going to setup SSO for.',
+      body: {
+        title: 'Select your identity provider',
+        description: "We'll guide you through the detailed setup process next.",
+      },
+      saml: {
+        groupLabel: 'SAML',
+        okta: 'Okta Workforce',
+        customSaml: 'Custom SAML Provider',
+      },
+      warning: 'Once a provider is selected you cannot change again until the configuration is over',
     },
     verifyEmailDomainStep: {
       title: 'Электрондық пошта мекенжайын растау',