+++ packages/localizations/src/ko-KR.ts
@@ -148,7 +148,7 @@ export const koKR: LocalizationResource = {
       },
     },
     paymentMethods__label: '결제 수단',
-    popular: '인기',
+    highlightedPlanBadge: '인기',
     pricingTable: {
       billingCycle: '결제 주기',
       included: '포함',
@@ -192,6 +192,20 @@ export const koKR: LocalizationResource = {
     },
     navbar: {
       title: '싱글 사인온(SSO) 구성',
+    },
+    selectProviderStep: {
+      title: '공급자 선택',
+      subtitle: 'SSO를 설정할 공급자를 선택하세요.',
+      body: {
+        title: 'ID 공급자를 선택하세요',
+        description: '다음 단계에서 자세한 설정 프로세스를 안내해 드립니다.',
+      },
+      saml: {
+        groupLabel: 'SAML',
+        okta: 'Okta Workforce',
+        customSaml: '사용자 지정 SAML 공급자',
+      },
+      warning: '공급자를 선택하면 구성이 완료될 때까지 다시 변경할 수 없습니다',
     },
     verifyEmailDomainStep: {
       title: '이메일 주소 확인',