+++ packages/localizations/src/sv-SE.ts
@@ -141,7 +141,7 @@ export const svSE: LocalizationResource = {
       },
     },
     paymentMethods__label: undefined,
-    popular: undefined,
+    highlightedPlanBadge: undefined,
     pricingTable: {
       billingCycle: undefined,
       included: undefined,
@@ -185,6 +185,20 @@ export const svSE: LocalizationResource = {
     },
     navbar: {
       title: 'Konfigurera enkel inloggning (SSO)',
+    },
+    selectProviderStep: {
+      title: 'Välj leverantör',
+      subtitle: 'Välj den leverantör du ska konfigurera SSO för.',
+      body: {
+        title: 'Välj din identitetsleverantör',
+        description: 'Vi guidar dig sedan genom den detaljerade konfigurationsprocessen.',
+      },
+      saml: {
+        groupLabel: 'SAML',
+        okta: 'Okta Workforce',
+        customSaml: 'Anpassad SAML-leverantör',
+      },
+      warning: 'När en leverantör har valts kan du inte ändra igen förrän konfigurationen är klar',
     },
     verifyEmailDomainStep: {
       title: 'Verifiera e-postadress',