+++ packages/ui/src/elements/LineItems.tsx
@@ -270,7 +270,7 @@ function CopyButton({ text, copyLabel = 'Copy' }: { text: string; copyLabel?: st
   return (
     <Button
       variant='unstyled'
-      onClick={onCopy}
+      onClick={() => onCopy()}
       sx={t => ({
         color: 'inherit',
         width: t.sizes.$4,